refactor: Student List Printout components
